od-unix: add package build compatibility defines
Add the Unix compatibility definitions needed by the packaged build
after the shared portability fixes are in place.

+++ od-unix/sysconfig.h
@@ -156,8 +156,8 @@ typedef int16_t INT16;
 typedef uint16_t UINT16;
 typedef int32_t INT32;
 typedef uint32_t UINT32;
-typedef int64_t INT64;
-typedef uint64_t UINT64;
+typedef signed long long INT64;
+typedef unsigned long long UINT64;
 #endif
 
 #define BYTE int8_t
